atualizações com o JWT

+++ src/controllers/reservaController.js
@@ -35,7 +35,7 @@ module.exports = class ReservaController {
       return res.status(400).json(erroValidacao);
     }
     if (Number(verificarToken) !== Number(fk_id_usuario)){
-      return res.status(400).json({message: "Você não pode fazer reserva com esse ID!"});
+      return res.status(400).json({message: "Você não pode fazer reserva com ID de outro usuário!"});
     }
 
     try {
@@ -109,17 +109,23 @@ module.exports = class ReservaController {
   }
 
   static async updateReserva(req, res) {
-    const { data, hora_inicio, hora_fim } = req.body;
+    const { fk_id_usuario, data, hora_inicio, hora_fim } = req.body;
     const reservaId = req.params.id_reserva;
+    const verificarToken =  req.userId;
 
     // Valida os campos de atualização
     const erroValidacao = validateReserva.validarCamposAtualizacao({
+      fk_id_usuario,
       data,
       hora_inicio,
       hora_fim,
     });
     if (erroValidacao) {
       return res.status(400).json(erroValidacao);
+    }
+
+    if (Number(verificarToken) !== Number(fk_id_usuario)){
+      return res.status(400).json({message: "Você não pode atualizar a reserva de outro usuário!"});
     }
 
     try {
@@ -170,6 +176,12 @@ module.exports = class ReservaController {
 
   static async deleteReserva(req, res) {
     const reservaId = req.params.id_reserva;
+    const usuarioId = req.params.id_usuario;
+    const verificarToken =  req.userId;
+
+    if (Number(verificarToken) !== Number(usuarioId)){
+      return res.status(400).json({message: "Você não pode deletar a reserva de outro usuário!"});
+    }
     const query = `DELETE FROM reserva WHERE id_reserva = ?`;
     try {
       const results = await queryAsync(query, [reservaId]);
